Make tests compatible with unittest discover
